requirements.txt
adal


aiohttp


aiosignal


alembic


anaconda-client


anaconda-navigator


anaconda-project

annotated-types


antlr4-python3-runtime


anyio


applicationinsights


appnope


argcomplete


argon2-cffi


argon2-cffi-bindings


arrow


asttokens


async-lru


async-timeout


attrs


azure-appconfiguration


azure-batch


azure-cli


azure-cli-core


azure-cli-telemetry


azure-common


azure-core


azure-cosmos


azure-data-tables


azure-datalake-store


azure-graphrbac


azure-identity


azure-keyvault


azure-keyvault-administration


azure-keyvault-certificates


azure-keyvault-keys


azure-keyvault-secrets


azure-loganalytics


azure-mgmt-advisor


azure-mgmt-apimanagement


azure-mgmt-appconfiguration


azure-mgmt-appcontainers


azure-mgmt-applicationinsights


azure-mgmt-authorization


azure-mgmt-batch


azure-mgmt-batchai


azure-mgmt-billing


azure-mgmt-botservice


azure-mgmt-cdn


azure-mgmt-cognitiveservices


azure-mgmt-compute


azure-mgmt-consumption


azure-mgmt-containerinstance


azure-mgmt-containerregistry


azure-mgmt-containerservice


azure-mgmt-core


azure-mgmt-cosmosdb


azure-mgmt-databoxedge


azure-mgmt-datalake-analytics


azure-mgmt-datalake-nspkg


azure-mgmt-datalake-store


azure-mgmt-datamigration


azure-mgmt-devtestlabs


azure-mgmt-dns


azure-mgmt-eventgrid


azure-mgmt-eventhub


azure-mgmt-extendedlocation


azure-mgmt-hdinsight


azure-mgmt-imagebuilder


azure-mgmt-iotcentral


azure-mgmt-iothub


azure-mgmt-iothubprovisioningservices


azure-mgmt-keyvault


azure-mgmt-kusto


azure-mgmt-loganalytics


azure-mgmt-managedservices


azure-mgmt-managementgroups


azure-mgmt-maps


azure-mgmt-marketplaceordering


azure-mgmt-media


azure-mgmt-monitor


azure-mgmt-msi


azure-mgmt-netapp


azure-mgmt-nspkg


azure-mgmt-policyinsights


azure-mgmt-privatedns


azure-mgmt-rdbms


azure-mgmt-recoveryservices


azure-mgmt-recoveryservicesbackup


azure-mgmt-redhatopenshift


azure-mgmt-redis


azure-mgmt-relay


azure-mgmt-resource


azure-mgmt-search


azure-mgmt-security


azure-mgmt-servicebus


azure-mgmt-servicefabric


azure-mgmt-servicefabricmanagedclusters


azure-mgmt-servicelinker


azure-mgmt-signalr


azure-mgmt-sql


azure-mgmt-sqlvirtualmachine


azure-mgmt-storage


azure-mgmt-synapse


azure-mgmt-trafficmanager


azure-mgmt-web


azure-multiapi-storage


azure-nspkg


azure-storage-blob


azure-storage-common


azure-synapse-accesscontrol


azure-synapse-artifacts


azure-synapse-managedprivateendpoints


azure-synapse-spark


azureml-core


azureml-mlflow


Babel


backcall


backoff


backports.functools-lru-cache

backports.tempfile

backports.weakref


bcrypt


beautifulsoup4

bidict


bleach


blinker


blis


boltons

brotlipy


catalogue


cattrs


certifi

cffi

chardet


charset-normalizer

click


cloudpickle


clyent


colorama


comm


conda

conda-build

conda-content-trust

conda-pack

conda-package-handling

conda-repo-cli


conda-token

conda-verify


conda_index

conda_package_streaming

confection


contextlib2


contourpy


cryptography

cycler


cymem


databricks-cli


datasets


debugpy


decorator


defusedxml

Deprecated


dill


distlib


docker


dsp-ml


dspy


dspy-ai


entrypoints


executing


fabric


fastjsonschema

filelock


Flask


fonttools


fqdn


frozenlist


fsspec


future

gitdb


GitPython


glob2

greenlet


gunicorn


huggingface-hub


humanfriendly


idna

importlib-metadata


invoke


ipykernel


ipython


ipython-genutils


ipywidgets


isodate


isoduration


itsdangerous


javaproperties


jedi


jeepney


Jinja2

jmespath


joblib


json5


jsondiff


jsonpatch

jsonpickle


jsonpointer


jsonschema


jsonschema-specifications


jupyter


jupyter-console


jupyter-events


jupyter-lsp


jupyter_client


jupyter_core

jupyter_server


jupyter_server_terminals


jupyterlab


jupyterlab-pygments


jupyterlab-widgets


jupyterlab_server


kiwisolver


knack


langcodes


libarchive-c

Mako


Markdown


MarkupSafe

matplotlib


matplotlib-inline


mistune


mlflow


mlflow-skinny


more-itertools

mpmath


msal


msal-extensions


msrest


msrestazure


multidict


multiprocess


murmurhash


navigator-updater


nbclient


nbconvert


nbformat

ndg-httpsclient


nest-asyncio


networkx


nltk


notebook


notebook_shim


numpy


oauthlib


openai


overrides


packaging

pandas


pandocfilters


paramiko


parso


pathlib

pathlib2


pathspec


pathy


pexpect


pickleshare


Pillow


pkginfo

platformdirs


playgroundrl


pluggy

ply


portalocker


preshed


prometheus-client


prompt-toolkit


protobuf


psutil

ptyprocess


pure-eval


pyarrow


pyasn1


pycosat

pycparser

pydantic


pydantic_core


PyGithub


Pygments


PyJWT

PyNaCl


pyOpenSSL

pyparsing


PyQt5-sip


pyrsistent

PySocks

python-dateutil

python-engineio


python-json-logger


python-socketio


pytz

PyYAML

pyzmq


qtconsole


QtPy

querystring-parser


referencing


regex


requests

requests-oauthlib


requests-toolbelt

rfc3339-validator


rfc3986-validator


rpds-py


ruamel-yaml-conda

ruamel.yaml

ruamel.yaml.clib

scikit-learn


scipy


scp


SecretStorage


semver


Send2Trash


sip

six

smart-open


smmap


sniffio


soupsieve

spacy


spacy-legacy


spacy-loggers


SQLAlchemy


sqlparse


srsly


sshtunnel


stack-data


sympy


tabulate


terminado


thinc


threadpoolctl


tinycss2


toml

toolz

torch


torchdata


torchtext


tornado

tqdm

traitlets

typer


typing_extensions


tzdata


ujson

uri-template


urllib3

virtualenv


wasabi


wcwidth


webcolors


webencodings


websocket-client


Werkzeug


widgetsnbextension


wrapt


xmltodict


xxhash


yarl


zipp


zstandard

